be/app/requirements.txt
uvicorn
fastapi
requests
pytz
icalendar
recurring-ical-events
pandas
python-dotenv
debugpy
ipykernel
odfpy
